Authorization bug fixed. Entities modified.

+++ Src/Application/OrderOperations.Application/DTOs/BasketDTOs/AddProductToBasketViewModel.cs
@@ -4,5 +4,5 @@ public class AddProductToBasketViewModel
 {
     public Guid BasketId { get; set; }
     public Guid ProductId { get; set; }
-    public double Quantity { get; set; } // Eklenecek ürün adedi
+    public decimal Quantity { get; set; } // Eklenecek ürün adedi
 }
