+++ Src/Application/OrderOperations.Application/DTOs/ProductDTOs/ProductStockUpdateViewModel.cs
@@ -3,5 +3,5 @@
 public class ProductStockUpdateViewModel
 {
     public Guid ProductId { get; set; }   // Stok eklenecek ürün ID'si
-    public double Quantity { get; set; }  // Eklenecek miktar
+    public decimal Quantity { get; set; }  // Eklenecek miktar
 }
