+++ Src/Domain/OrderPoerations.Domain/Entities/BasketItem.cs
@@ -5,6 +5,6 @@ namespace OrderPoerations.Domain.Entities;
 public class BasketItem : BaseAuditableEntity<int, Guid>
 {
     public Product Product { get; set; }
-    public double Amount { get; set; }
-    public double TotalCost { get; set; }
+    public decimal Amount { get; set; }
+    public decimal TotalCost { get; set; }
 }
